# improve performance of cpf algorithm