--- a/.gitignore
+++ b/.gitignore
@@ -7,3 +7,4 @@
 *.out
 .DS_Store
 go.work
+*.txt