--- a/cpf.go
+++ b/cpf.go
@@ -3,8 +3,6 @@ package br
 import (
 	"database/sql/driver"
 	"errors"
-
-	"github.com/phenpessoa/gutils/unsafex"
 )
 
 var (
@@ -32,62 +30,180 @@ type CPF string
 // IsValid checks whether the provided CPF is valid based on its checksum
 // digits.
 func (cpf CPF) IsValid() bool {
-	l := len(cpf)
-	if l != 11 && l != 14 {
+	switch len(cpf) {
+	case 11:
+		dByte, ok := iterFirst11(cpf)
+		if !ok {
+			return false
+		}
+
+		if cpf[len(cpf)-2] != dByte {
+			return false
+		}
+
+		dByte, ok = iterSecond11(cpf)
+		if !ok {
+			return false
+		}
+
+		return cpf[len(cpf)-1] == dByte
+	case 14:
+		if cpf[3] != '.' || cpf[7] != '.' || cpf[11] != '-' {
+			return false
+		}
+
+		dByte, ok := iterFirst14(cpf)
+		if !ok {
+			return false
+		}
+
+		if cpf[len(cpf)-2] != dByte {
+			return false
+		}
+
+		dByte, ok = iterSecond14(cpf)
+		if !ok {
+			return false
+		}
+
+		return cpf[len(cpf)-1] == dByte
+	default:
 		return false
 	}
-
-	dByte, ok := iterCPFTable(cpf, cpfFirstTable)
-	if !ok {
-		return false
-	}
-
-	if cpf[l-2] != dByte {
-		return false
-	}
-
-	dByte, ok = iterCPFTable(cpf, cpfSecondTable)
-	if !ok {
-		return false
-	}
-
-	return cpf[l-1] == dByte
 }
 
-func iterCPFTable(cpf CPF, table []int) (byte, bool) {
-	var sum, pad, rest, d int
+func iterFirst14(cpf CPF) (byte, bool) {
+	if len(cpf) != 14 || len(cpfFirstTable) != 9 {
+		panic("not 14 or 9")
+	}
 
-	for i, d := range table {
-		cur := cpf[i+pad]
-		switch cur {
-		case '.':
-			if i != 3 && i != 6 {
-				return 0, false
-			}
-			pad++
-			cur = cpf[i+pad]
-		case '-':
-			if i != 9 {
-				return 0, false
-			}
-			pad++
-			cur = cpf[i+pad]
-		}
+	var sum, rest, out int
 
-		if cur < '0' || cur > '9' {
+	for i, d := range cpfFirstTable[:3] {
+		cur := cpf[i]
+		if !isDigit(cur) {
 			return 0, false
 		}
+		sum += d * int(cur-'0')
+	}
 
+	for i, d := range cpfFirstTable[3:6] {
+		cur := cpf[4:7][i]
+		if !isDigit(cur) {
+			return 0, false
+		}
+		sum += d * int(cur-'0')
+	}
+
+	for i, d := range cpfFirstTable[6:9] {
+		cur := cpf[8:11][i]
+		if !isDigit(cur) {
+			return 0, false
+		}
 		sum += d * int(cur-'0')
 	}
 
 	rest = sum % 11
 
 	if rest >= 2 {
-		d = 11 - rest
+		out = 11 - rest
 	}
 
-	return byte(d) + '0', true
+	return byte(out) + '0', true
+}
+
+func iterSecond14(cpf CPF) (byte, bool) {
+	if len(cpf) != 14 || len(cpfSecondTable) != 10 {
+		panic("not 14 or 10")
+	}
+
+	var sum, rest, out int
+
+	for i, d := range cpfSecondTable[:3] {
+		cur := cpf[i]
+		if !isDigit(cur) {
+			return 0, false
+		}
+		sum += d * int(cur-'0')
+	}
+
+	for i, d := range cpfSecondTable[3:6] {
+		cur := cpf[4:7][i]
+		if !isDigit(cur) {
+			return 0, false
+		}
+		sum += d * int(cur-'0')
+	}
+
+	for i, d := range cpfSecondTable[6:9] {
+		cur := cpf[8:11][i]
+		if !isDigit(cur) {
+			return 0, false
+		}
+		sum += d * int(cur-'0')
+	}
+
+	last := cpf[12]
+	if !isDigit(last) {
+		return 0, false
+	}
+	sum += cpfSecondTable[9] * int(last-'0')
+
+	rest = sum % 11
+
+	if rest >= 2 {
+		out = 11 - rest
+	}
+
+	return byte(out) + '0', true
+}
+
+func iterFirst11(cpf CPF) (byte, bool) {
+	if len(cpf) != 11 || len(cpfFirstTable) != 9 {
+		panic("not 11 or 9")
+	}
+
+	var sum, rest, out int
+
+	for i, d := range cpfFirstTable {
+		cur := cpf[i]
+		if !isDigit(cur) {
+			return 0, false
+		}
+		sum += d * int(cur-'0')
+	}
+
+	rest = sum % 11
+
+	if rest >= 2 {
+		out = 11 - rest
+	}
+
+	return byte(out) + '0', true
+}
+
+func iterSecond11(cpf CPF) (byte, bool) {
+	if len(cpf) != 11 || len(cpfSecondTable) != 10 {
+		panic("not 11 or 10")
+	}
+
+	var sum, rest, out int
+
+	for i, d := range cpfSecondTable {
+		cur := cpf[i]
+		if !isDigit(cur) {
+			return 0, false
+		}
+		sum += d * int(cur-'0')
+	}
+
+	rest = sum % 11
+
+	if rest >= 2 {
+		out = 11 - rest
+	}
+
+	return byte(out) + '0', true
 }
 
 // String returns the formatted CPF string with punctuation as XXX.XXX.XXX-XX.
@@ -100,30 +216,22 @@ func (cpf CPF) String() string {
 		return string(cpf)
 	}
 
-	out := make([]byte, 14)
-	for i := range cpf {
-		switch {
-		case i < 3:
-			out[i] = cpf[i]
-		case i == 3:
-			out[i] = '.'
-			out[i+1] = cpf[i]
-		case i < 6:
-			out[i+1] = cpf[i]
-		case i == 6:
-			out[i+1] = '.'
-			out[i+2] = cpf[i]
-		case i < 9:
-			out[i+2] = cpf[i]
-		case i == 9:
-			out[i+2] = '-'
-			out[i+3] = cpf[i]
-		default:
-			out[i+3] = cpf[i]
-		}
+	if len(cpf) != 11 {
+		panic("not 11")
 	}
 
-	return unsafex.String(out)
+	out := make([]byte, 14)
+
+	out[3] = '.'
+	out[7] = '.'
+	out[11] = '-'
+
+	copy(out[0:3], cpf[0:3])
+	copy(out[4:7], cpf[3:6])
+	copy(out[8:11], cpf[6:9])
+	copy(out[12:14], cpf[9:11])
+
+	return string(out)
 }
 
 func (c CPF) Value() (driver.Value, error) {