--- a/utils_test.go
+++ b/utils_test.go
@@ -1,0 +1,6 @@
+package br
+
+var (
+	boolSink   bool
+	stringSink string
+)