```
goos: linux
goarch: amd64
pkg: github.com/phenpessoa/br
cpu: Intel(R) Core(TM) Ultra 7 155H
                        │   old.txt   │               new.txt               │
                        │   sec/op    │   sec/op     vs base                │
CPF_IsValid14-21          26.88n ± 1%   15.01n ± 2%  -44.15% (p=0.000 n=10)
CPF_IsValid11-21          21.53n ± 2%   13.98n ± 3%  -35.05% (p=0.000 n=10)
CPF_IsValid14Invalid-21   26.57n ± 2%   14.93n ± 1%  -43.82% (p=0.000 n=10)
CPF_IsValid11Invalid-21   21.77n ± 2%   14.15n ± 2%  -34.98% (p=0.000 n=10)
CPF_String14-21           26.91n ± 1%   15.99n ± 2%  -40.56% (p=0.000 n=10)
CPF_String11-21           50.65n ± 1%   27.89n ± 2%  -44.94% (p=0.000 n=10)
geomean                   27.75n        16.45n       -40.73%
```